+++ packages/media-tracks/src/mixin.ts
@@ -123,7 +123,7 @@ export function MediaTracksMixin<T>(MediaElementClass: T): WithMediaTracks<T> {
     let renditions = getPrivate(media).videoRenditions;
     if (!renditions) {
       renditions = new VideoRenditionList();
-      getPrivate(renditions).media = media;
+      getPrivate(renditions).media = new WeakRef(media);
       getPrivate(media).videoRenditions = renditions;
     }
     return renditions;
@@ -141,7 +141,7 @@ export function MediaTracksMixin<T>(MediaElementClass: T): WithMediaTracks<T> {
     let renditions = getPrivate(media).audioRenditions;
     if (!renditions) {
       renditions = new AudioRenditionList();
-      getPrivate(renditions).media = media;
+      getPrivate(renditions).media = new WeakRef(media);
       getPrivate(media).audioRenditions = renditions;
     }
     return renditions;
@@ -171,11 +171,11 @@ function getVideoTracks(media: any) {
         addVideoTrack(media, nativeTrack);
       }
 
-      nativeTracks.addEventListener('change', () => {
+      const onChange = () => {
         tracks.dispatchEvent(new Event('change'));
-      });
+      };
 
-      nativeTracks.addEventListener('addtrack', (event: TrackEvent) => {
+      const onAddTrack = (event: TrackEvent) => {
         // Note: adding native track instances to the shim track list here.
         // This works because the API is identical and change event is forwarded.
         // If tracks were manually added prevent native tracks from being added...
@@ -189,11 +189,15 @@ function getVideoTracks(media: any) {
         }
 
         addVideoTrack(media, event.track as VideoTrack);
-      });
+      };
 
-      nativeTracks.addEventListener('removetrack', (event: TrackEvent) => {
-        removeVideoTrack(event.track  as VideoTrack);
-      });
+      const onRemoveTrack = (event: TrackEvent) => {
+        removeVideoTrack(event.track as VideoTrack);
+      };
+
+      nativeTracks.addEventListener('change', onChange);
+      nativeTracks.addEventListener('addtrack', onAddTrack);
+      nativeTracks.addEventListener('removetrack', onRemoveTrack);
     }
   }
   return tracks;
@@ -214,11 +218,11 @@ function getAudioTracks(media: any) {
         addAudioTrack(media, nativeTrack);
       }
 
-      nativeTracks.addEventListener('change', () => {
+      const onChange = () => {
         tracks.dispatchEvent(new Event('change'));
-      });
+      };
 
-      nativeTracks.addEventListener('addtrack', (event: TrackEvent) => {
+      const onAddTrack = (event: TrackEvent) => {
         // Note: adding native track instances to the shim track list here.
         // This works because the API is identical and change event is forwarded.
         // If tracks were manually added prevent native tracks from being added...
@@ -232,11 +236,15 @@ function getAudioTracks(media: any) {
         }
 
         addAudioTrack(media, event.track as AudioTrack);
-      });
+      };
 
-      nativeTracks.addEventListener('removetrack', (event: TrackEvent) => {
+      const onRemoveTrack = (event: TrackEvent) => {
         removeAudioTrack(event.track as AudioTrack);
-      });
+      };
+
+      nativeTracks.addEventListener('change', onChange);
+      nativeTracks.addEventListener('addtrack', onAddTrack);
+      nativeTracks.addEventListener('removetrack', onRemoveTrack);
     }
   }
   return tracks;